requirements.txt
Django
psycopg2
djangorestframework
django-filter
markdown
gremlinpython
requests